Move admin bar handler to Interactivity API
Replace the inline DOM click listener in src/frontend.js with a dedicated interactivity module. Adds src/interactivity/admin-bar.js which registers a p2026/admin-bar store action (handleDocumentClick) wired to document clicks via a hidden interactivity host. The new action preserves behavior: if a new-post editor exists, scroll and focus it; otherwise dispatch openNewPostModal(). frontend.js now imports and calls initAdminBarInteractivity() and removes the previous inline handler and some now-unused imports.
